Pull up PushDownAndCombineFilters
